@@ -1,11 +1,18 @@
 import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
-import { UntypedFormControl, Validators } from '@angular/forms';
+import { UntypedFormControl, Validators, ReactiveFormsModule } from '@angular/forms';
 import { Topic } from 'src/app/services/pubsub.service';
+import { MatButton } from '@angular/material/button';
+import { MatIcon } from '@angular/material/icon';
+import { MatFormField } from '@angular/material/form-field';
+import { MatInput } from '@angular/material/input';
+import { CdkTextareaAutosize } from '@angular/cdk/text-field';
 
 @Component({
-  selector: 'app-topic-details',
-  templateUrl: './topic-details.component.html',
-  styleUrls: ['./topic-details.component.scss']
+    selector: 'app-topic-details',
+    templateUrl: './topic-details.component.html',
+    styleUrls: ['./topic-details.component.scss'],
+    standalone: true,
+    imports: [MatButton, MatIcon, MatFormField, MatInput, CdkTextareaAutosize, ReactiveFormsModule]
 })
 export class TopicDetailsComponent implements OnInit {
 
