@@ -1,11 +1,11 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import { IndexComponent } from './components/index/index.component';
-import { ProjectsComponent } from './components/projects/projects.component';
+
+
 
 const routes: Routes = [
-  {path: '', component: IndexComponent},
-  {path: 'project', component: ProjectsComponent},
+  {path: '', loadComponent: () => import('./components/index/index.component').then(m => m.IndexComponent)},
+  {path: 'project', loadComponent: () => import('./components/projects/projects.component').then(m => m.ProjectsComponent)},
 ];
 
 @NgModule({
