@@ -1,19 +1,30 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, OnInit, inject } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { Observable, filter } from 'rxjs';
 import { PubsubService } from 'src/app/services/pubsub.service';
 import { InputDialogComponent } from '../input-dialog/input-dialog.component';
+import { AsyncPipe } from '@angular/common';
+import { MatButton } from '@angular/material/button';
+import { RouterLink } from '@angular/router';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
-  selector: 'app-index',
-  templateUrl: './index.component.html',
-  styleUrls: ['./index.component.scss']
+    selector: 'app-index',
+    templateUrl: './index.component.html',
+    styleUrls: ['./index.component.scss'],
+    standalone: true,
+    imports: [MatButton, RouterLink, MatIcon, AsyncPipe]
 })
 export class IndexComponent implements OnInit {
+  private pubsub = inject(PubsubService);
+  private matDialog = inject(MatDialog);
+
 
   projectList$: Observable<string[]>
 
-  constructor(private pubsub: PubsubService, private matDialog: MatDialog) {
+  constructor() {
+    const pubsub = this.pubsub;
+
     this.projectList$ = pubsub.projectList$
   }
 
