@@ -1,15 +1,22 @@
-import { Component, EventEmitter, Input, OnInit, Output } from '@angular/core';
+import { Component, EventEmitter, Input, OnInit, Output, inject } from '@angular/core';
 import { MatDialog } from '@angular/material/dialog';
 import { firstValueFrom } from 'rxjs';
 import { Subscription, Topic } from 'src/app/services/pubsub.service';
 import { NewSubscriptionDialogComponent, NewSubscriptionRequest } from './new-subscription-dialog/new-subscription-dialog.component';
+import { MatActionList, MatListItem } from '@angular/material/list';
+import { NgClass } from '@angular/common';
+import { MatIcon } from '@angular/material/icon';
 
 @Component({
-  selector: 'app-subscription-list',
-  templateUrl: './subscription-list.component.html',
-  styleUrls: ['./subscription-list.component.scss']
+    selector: 'app-subscription-list',
+    templateUrl: './subscription-list.component.html',
+    styleUrls: ['./subscription-list.component.scss'],
+    standalone: true,
+    imports: [MatActionList, MatListItem, NgClass, MatIcon]
 })
 export class SubscriptionListComponent implements OnInit {
+  private dialog = inject(MatDialog);
+
 
   @Input() subscriptions?: Subscription[]
   @Input() topic?: Topic
@@ -17,7 +24,6 @@ export class SubscriptionListComponent implements OnInit {
   @Input() currentSubscription?: Subscription
   @Output() currentSubscriptionChange = new EventEmitter<Subscription>()
   @Output() newSubscriptionRequest = new EventEmitter<NewSubscriptionRequest>()
-  constructor(private dialog: MatDialog) { }
 
   ngOnInit(): void {
 
